requirements.txt
yamlinfo
python-dotenv
elasticsearch
requests
ujson
click
udatetime
DataParserLib